prevent double likes coming through

+++ src/routes/api/posts/[postId]/comments/[commentId]/unlike/+server.ts
@@ -23,7 +23,19 @@ export const POST = async ({ locals, params }: RequestEvent) => {
 			throw error(400, 'Invalid Comment ID');
 		}
 
-		// Like comment
+		// Check if user has liked the comment
+		const existingLike = await db.like.findFirst({
+			where: {
+				userId: locals.user.id,
+				commentId: commentId
+			}
+		});
+
+		if (!existingLike) {
+			throw error(400, 'User has not liked this comment');
+		}
+
+		// Unlike comment
 		await db.like.deleteMany({
 			where: {
 				userId: locals.user.id,
